COMP: warnings.
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=8218
git-svn-id: http://svn.slicer.org/Slicer4/trunk@8218 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/ChangeTracker/vtkChangeTrackerLogic.cxx
@@ -758,7 +758,10 @@ vtkImageData*  vtkChangeTrackerLogic::GetAnalysis_Intensity_ROIBinDisplay() {
 
 void vtkChangeTrackerLogic::MeassureGrowth(double &Shrinkage, double &Growth) {
    if (this->ChangeTrackerNode) {
-     this->MeassureGrowth(this->ChangeTrackerNode->GetSegmentThresholdMin(),this->ChangeTrackerNode->GetSegmentThresholdMax(),Shrinkage,Growth); 
+   this->MeassureGrowth(
+     static_cast<int>(this->ChangeTrackerNode->GetSegmentThresholdMin()),
+     static_cast<int>(this->ChangeTrackerNode->GetSegmentThresholdMax()),
+     Shrinkage, Growth); 
    }  else {
      cout << "Error: vtkChangeTrackerLogic::MeassureGrowth: No ChangeTrackerNode defined" << endl;
    }
